Tuning DraftDigest
